Extraction of model elements visit code (and pom decorator)
